Practice23 (Spring 14)/src/main/java/com/example/practice23/services/UserService.java
package com.example.practice23.services;

import com.example.practice23.models.User;
import com.example.practice23.models.enums.Role;
import com.example.practice23.repositories.UserRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
@Slf4j
@RequiredArgsConstructor
public class UserService {
    private final UserRepository userRepository;
    private final PasswordEncoder passwordEncoder;

    public boolean createUser(User user) {
        final String email = user.getEmail();
        if (userRepository.findByEmail(user.getEmail()) != null) {
            log.info("User with email {} already exists", email);
            return false;
        }

        user.setActive(true);
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        user.getRoles().add(Role.ROLE_USER);
        log.info("Saving new user with email: {}", email);
        userRepository.save(user);
        return true;
    }
}
